+++ enterprise/frontend/src/metabase-enterprise/dependencies/components/DependencyGraph/GraphDependencyPanel/utils.ts
@@ -1,15 +1,13 @@
 import type {
+  DependencyFilterOptions,
   DependencyGroupType,
   DependencyNode,
   DependencySortColumn,
+  DependencySortOptions,
   ListNodeDependentsRequest,
 } from "metabase-types/api";
 
 import { DEFAULT_INCLUDE_PERSONAL_COLLECTIONS } from "../../../constants";
-import type {
-  DependencyFilterOptions,
-  DependencySortOptions,
-} from "../../../types";
 import {
   canHaveViewCount,
   getCardType,
