+++ enterprise/frontend/src/metabase-enterprise/dependencies/components/FilterOptionsPicker/LocationFilterPicker/LocationFilterPicker.tsx
@@ -2,8 +2,7 @@ import type { ChangeEvent } from "react";
 import { t } from "ttag";
 
 import { Checkbox, Input, Stack } from "metabase/ui";
-
-import type { DependencyFilterOptions } from "../../../types";
+import type { DependencyFilterOptions } from "metabase-types/api";
 
 type LocationFilterPickerProps = {
   filterOptions: DependencyFilterOptions;
