+++ enterprise/frontend/src/metabase-enterprise/dependencies/components/FilterOptionsPicker/TypeFilterPicker/TypeFilterPicker.tsx
@@ -2,9 +2,11 @@ import { useState } from "react";
 import { t } from "ttag";
 
 import { Checkbox, Stack } from "metabase/ui";
-import type { DependencyGroupType } from "metabase-types/api";
+import type {
+  DependencyFilterOptions,
+  DependencyGroupType,
+} from "metabase-types/api";
 
-import type { DependencyFilterOptions } from "../../../types";
 import { getDependencyGroupTypeInfo } from "../../../utils";
 
 type TypeFilterPickerProps = {
