+++ enterprise/frontend/src/metabase-enterprise/dependencies/utils.ts
@@ -8,6 +8,7 @@ import type {
   AnalysisFindingErrorType,
   CardType,
   DependencyEntry,
+  DependencyFilterOptions,
   DependencyGroupType,
   DependencyId,
   DependencyNode,
@@ -22,7 +23,6 @@ import type {
   DependencyError,
   DependencyErrorGroup,
   DependencyErrorInfo,
-  DependencyFilterOptions,
   DependencyGroupTypeInfo,
   DependentGroup,
   NodeId,
